+++ inst/htmlwidgets/RacViewer/lib/functions/load.js
@@ -63,9 +63,22 @@ Racmacs.Viewer.prototype.load = function(
         // Generate the antigen and sera objects
         this.addAntigensAndSera();
 
+        // Get plot limits
+        if (!options.xlim) var xlim = this.data.xlim();
+        else               var xlim = options.xlim;
+        if (!options.ylim) var ylim = this.data.ylim();
+        else               var ylim = options.ylim;
+        var zlim = this.data.zlim();
+        console.log(xlim);
+        console.log(ylim);
+
         // Set dims from map data
         if(!options.maintain_viewpoint){
-            this.setDims();
+            this.setDims({
+                xlim : xlim,
+                ylim : ylim,
+                zlim : zlim
+            });
         }
 
         // Load the hi table
@@ -80,9 +93,6 @@ Racmacs.Viewer.prototype.load = function(
         if(this.data.numProjections() > 0) {
             this.addAgSrPoints();
         }
-
-        // Update the stress
-        this.updateStress(this.data.stress());
 
         // Update antigen and sera browsers
         this.browsers.antigens.populate();
@@ -131,14 +141,18 @@ Racmacs.Viewer.prototype.load = function(
             var zoom = this.data.getViewerSetting("zoom");
             if(zoom === null){
                 
-                var xlim = this.data.xlim();
-                var ylim = this.data.ylim();
-                var zlim = this.data.zlim();
-
                 var scale = this.scene.getWorldScale();
                 xlim = [xlim[0]*scale[0], xlim[1]*scale[0]];
                 ylim = [ylim[0]*scale[1], ylim[1]*scale[1]];
                 zlim = [zlim[0]*scale[1], zlim[1]*scale[1]];
+
+                var translation = [
+                  (xlim[1] + xlim[0]) / 2,
+                  (ylim[1] + ylim[0]) / 2,
+                  (zlim[1] + zlim[0]) / 2
+                ];
+
+                // this.scene.setTranslation(translation);
 
                 this.camera.zoomToLims({
                     x : xlim,
@@ -194,6 +208,12 @@ Racmacs.Viewer.prototype.load = function(
         if(options["show.connectionlines"])  this.showConnectionLines();
         if(options["show.errorlines"])       this.showErrorLines();
         if(options["show.titers"])           this.showTiters();
+
+        // Update the viewer frustrum
+        this.updateFrustrum();
+
+        // Update the stress
+        this.updateStress(this.data.stress());
 
         // Note that content is now loaded
         this.contentLoaded = true;