+++ NEWS.md
@@ -28,3 +28,6 @@
 * Add option in `plot.acmap()` to draw stress lines
 * Add xlim and ylim options to `view.acmap()` to control how maps are displayed
 * Add check to procrustes that there are sufficient matching points
+
+# Racmacs 1.1.13
+